+++ cognite/extractorutils/uploader/events.py
@@ -1,0 +1,165 @@
+#  Copyright 2023 Cognite AS
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+import threading
+from typing import Callable, List, Optional
+
+import arrow
+from cognite.client import CogniteClient
+from cognite.client.data_classes import Event
+from cognite.client.exceptions import CogniteAPIError
+from requests import ConnectionError
+
+from cognite.extractorutils.retry import retry
+from cognite.extractorutils.uploader._base import (
+    RETRIES,
+    RETRY_BACKOFF_FACTOR,
+    RETRY_DELAY,
+    RETRY_MAX_DELAY,
+    AbstractUploadQueue,
+)
+from cognite.extractorutils.uploader._metrics import (
+    EVENTS_UPLOADER_LATENCY,
+    EVENTS_UPLOADER_QUEUE_SIZE,
+    EVENTS_UPLOADER_QUEUED,
+    EVENTS_UPLOADER_WRITTEN,
+)
+
+
+class EventUploadQueue(AbstractUploadQueue):
+    """
+    Upload queue for events
+
+    Args:
+        cdf_client: Cognite Data Fusion client to use
+        post_upload_function: A function that will be called after each upload. The function will be given one argument:
+            A list of the events that were uploaded.
+        max_queue_size: Maximum size of upload queue. Defaults to no max size.
+        max_upload_interval: Automatically trigger an upload each m seconds when run as a thread (use start/stop
+            methods).
+        trigger_log_level: Log level to log upload triggers to.
+        thread_name: Thread name of uploader thread.
+    """
+
+    def __init__(
+        self,
+        cdf_client: CogniteClient,
+        post_upload_function: Optional[Callable[[List[Event]], None]] = None,
+        max_queue_size: Optional[int] = None,
+        max_upload_interval: Optional[int] = None,
+        trigger_log_level: str = "DEBUG",
+        thread_name: Optional[str] = None,
+        cancellation_token: threading.Event = threading.Event(),
+    ):
+        # Super sets post_upload and threshold
+        super().__init__(
+            cdf_client,
+            post_upload_function,
+            max_queue_size,
+            max_upload_interval,
+            trigger_log_level,
+            thread_name,
+            cancellation_token,
+        )
+
+        self.upload_queue: List[Event] = []
+
+        self.events_queued = EVENTS_UPLOADER_QUEUED
+        self.events_written = EVENTS_UPLOADER_WRITTEN
+        self.queue_size = EVENTS_UPLOADER_QUEUE_SIZE
+        self.latency = EVENTS_UPLOADER_LATENCY
+        self.latency_zero_point = arrow.utcnow()
+
+    def add_to_upload_queue(self, event: Event) -> None:
+        """
+        Add event to upload queue. The queue will be uploaded if the queue size is larger than the threshold
+        specified in the __init__.
+
+        Args:
+            event: Event to add
+        """
+        with self.lock:
+            if self.upload_queue_size == 0:
+                self.latency_zero_point = arrow.utcnow()
+
+            self.upload_queue.append(event)
+            self.events_queued.inc()
+            self.upload_queue_size += 1
+            self.queue_size.set(self.upload_queue_size)
+
+            self._check_triggers()
+
+    def upload(self) -> None:
+        """
+        Trigger an upload of the queue, clears queue afterwards
+        """
+        if len(self.upload_queue) == 0:
+            return
+
+        with self.lock:
+            self._upload_batch()
+
+            self.latency.observe(
+                (arrow.utcnow() - self.latency_zero_point).total_seconds() / 60
+            )  # show data in minutes
+            self.events_written.inc(self.upload_queue_size)
+
+            try:
+                self._post_upload(self.upload_queue)
+            except Exception as e:
+                self.logger.error("Error in upload callback: %s", str(e))
+            self.upload_queue.clear()
+            self.logger.info(f"Uploaded {self.upload_queue_size} events")
+            self.upload_queue_size = 0
+            self.queue_size.set(self.upload_queue_size)
+
+    @retry(
+        exceptions=(CogniteAPIError, ConnectionError),
+        tries=RETRIES,
+        delay=RETRY_DELAY,
+        max_delay=RETRY_MAX_DELAY,
+        backoff=RETRY_BACKOFF_FACTOR,
+    )
+    def _upload_batch(self):
+        self.cdf_client.events.create([e for e in self.upload_queue])
+
+    def __enter__(self) -> "EventUploadQueue":
+        """
+        Wraps around start method, for use as context manager
+
+        Returns:
+            self
+        """
+        self.start()
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb) -> None:
+        """
+        Wraps around stop method, for use as context manager
+
+        Args:
+            exc_type: Exception type
+            exc_val: Exception value
+            exc_tb: Traceback
+        """
+        self.stop()
+
+    def __len__(self) -> int:
+        """
+        The size of the upload queue
+
+        Returns:
+            Number of events in queue
+        """
+        return self.upload_queue_size